+++ client/src/commonMain/kotlin/eu/tiducto/spider/contract/routing/PlanConnectionStreamVariables.kt
@@ -3,14 +3,15 @@ package eu.tiducto.spider.contract.routing
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class PlanConnectionVariables(
+internal data class PlanConnectionStreamVariables(
     val dateTime: PlanDateTimeInput,
     val origin: PlanLabeledLocationInput,
     val destination: PlanLabeledLocationInput,
-    val searchWindow: String,
     val via: List<PlanViaLocationInput>? = null,
     val modes: PlanModesInput? = null,
     val preferences: PlanPreferencesInput? = null,
+    val targetResults: Int? = null,
+    val maxWindow: String? = null,
     val before: String? = null,
     val after: String? = null
 )